Don't check codes to be in range, but find "best match"
If some new events are added to an extension, older implementations register a
smaller portion of the event number space than this library possible knows. So
we cannot just look for some extension that has some code "in range", but we
need to find the best one (with the largest base).

Signed-off-by: Uli Schlachter <psychon@znc.in>